Change all time vars to uint64_t

+++ control/experiment/PID_experiment.c
@@ -2,6 +2,7 @@
 #include "virtual_intellisat.h"
 
 #include <math.h>
+#include <stdint.h>
 
 //TODO: IMU, HDD alternation?
 #define IMU_CHOICE VI_IMU1
@@ -15,7 +16,7 @@ PID_status PID_experiment()
         return PID_EXPERIMENT_FAILURE;
 
     //Get the current time (Virtual Intellisat)
-    int curr_millis = 0;
+    uint64_t curr_millis = 0;
     if(vi_get_curr_millis(&curr_millis) == GET_CURR_MILLIS_FAILURE)
         return PID_EXPERIMENT_FAILURE;
